+++ packages/vault_iota/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+- BREAKING CHANGE: ConsentStorage implementers must add deleteByHash.
+- Add IotaConsentRecordService.deleteConsentRecord.
+
 ## 1.0.0-dev
 
 - Initial version.